app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

httpx==0.27.2
python-dotenv==1.0.1

pyjwt[crypto]==2.9.0
cryptography==43.0.1

pydantic==2.8.2

qdrant-client==1.11.1

openai==1.40.0

prometheus-client==0.20.0